text writting

+++ graphics2/untitled1/glview.hpp
@@ -30,16 +30,20 @@ public:
 
 private:
     unsigned int calculatedPoints = 1;
+
     std::vector<QPoint> dotBuf; //user's points
     std::vector<QPoint> dotDrawingBuf; //points of spline
     QPoint mPosition; //mouse pos
+
     QTimer mTimer;
     bool UpdateOn = false;
+
     double mWindowScale = 1.0; //scaling
-    float mShiftX = 0; //moving by wasd
-    float mShiftY = 0; //moving by wasd
-    float mScaleFactorX; //window resizing
-    float mScaleFactorY; //window resizing
+    int mShiftX = 0; //moving by wasd
+    int mShiftY = 0; //moving by wasd
+
+    double mScaleFactorX; //window resizing
+    double mScaleFactorY; //window resizing
 
     void extendPoints(const int mul);
 };
